Use layer compression workarounds as fallback only
The compressed layers actually do match the original blobs in the case
where the blobs were originally built with the containers/storage tools
(e.g. buildah) after the point where the compression library was
changed. So try this in the first instance.

Layers built by other tools (e.g. moby) use the gzip algorithm from the
stdlib, so fall back to that if the built-in compression does not
produce a blob with the correct digest.

It nevertheless remains possible that other tools (or other versions of
tools) use different gzip implementations that we will not be able to
replicate.